reactor tests refactor
