+++ node/src/reactor/main_reactor/tests/initial_stakes.rs
@@ -1,0 +1,5 @@
+pub(crate) enum InitialStakes {
+    FromVec(Vec<u128>),
+    Random { count: usize },
+    AllEqual { count: usize, stake: u128 },
+}
